requirements.txt
fastapi
uvicorn
motor
dnspython
python-dotenv
aiofiles
